requirements.txt
# Requires Python 3.10+
torch==2.7.1+cu128
optuna==4.3.0
wandb==0.20.1
python-dotenv==1.1.0
PyYAML==6.0.2
pandas==2.3.0
numpy==1.26.4
scipy==1.15.3
matplotlib==3.10.3
gymnasium==1.1.1
ta==0.11.0
polygon-api-client==1.14.6
